# tiendascotia/src/main/java/com/bd_tienda_test/dto/FiltroDetalle.java
package com.bd_tienda_test.dto;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FiltroDetalle {

	private String parametro;

	private String valor;
}