# tiendascotia/src/main/java/com/bd_tienda_test/dto/FiltrosDto.java
package com.bd_tienda_test.dto;
import java.util.List;
import java.util.Map;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FiltrosDto {

	private List<FiltroDetalle> filtros;
}